Without observer, but updating session
(cherry picked from commit cf23c3db551d1d903196ff02ebf2e45f6c9e18eb)
