+++ app/src/main/java/io/lunarlogic/aircasting/models/SessionsViewModel.kt
@@ -15,6 +15,10 @@ class SessionsViewModel(): ViewModel() {
         return mDatabase.sessions().loadLiveDataSessionAndMeasurementsByUUID(uuid)
     }
 
+    fun reloadSession(uuid: String): SessionWithStreamsDBObject? {
+        return mDatabase.sessions().reloadSessionAndStreamsByUUID(uuid)
+    }
+
     fun reloadSessionWithMeasurements(uuid: String): SessionWithStreamsAndMeasurementsDBObject? {
         return mDatabase.sessions().reloadSessionAndMeasurementsByUUID(uuid)
     }
