+++ app/src/main/java/io/lunarlogic/aircasting/screens/session_view/SessionDetailsViewController.kt
@@ -6,10 +6,8 @@ import io.lunarlogic.aircasting.events.NewMeasurementEvent
 import io.lunarlogic.aircasting.lib.safeRegister
 import io.lunarlogic.aircasting.location.LocationHelper
 import io.lunarlogic.aircasting.models.*
-import io.lunarlogic.aircasting.models.observers.SessionObserver
 import io.lunarlogic.aircasting.screens.session_view.hlu.HLUValidationErrorToast
 import io.lunarlogic.aircasting.screens.dashboard.SessionPresenter
-import kotlinx.coroutines.CoroutineScope
 import org.greenrobot.eventbus.EventBus
 import org.greenrobot.eventbus.Subscribe
 import org.greenrobot.eventbus.ThreadMode
@@ -23,28 +21,45 @@ abstract class SessionDetailsViewController(
     private var sensorName: String?
 ): SessionDetailsViewMvc.Listener {
     private var mSessionPresenter = SessionPresenter(sessionUUID, sensorName)
-    private val mSessionObserver = SessionObserver(rootActivity, mSessionsViewModel, mSessionPresenter, this::onSessionChanged)
 
     fun onCreate() {
         EventBus.getDefault().safeRegister(this);
         mViewMvc?.registerListener(this)
-
-        mSessionObserver.observe()
-    }
-
-    private fun onSessionChanged(coroutineScope: CoroutineScope) {
-        DatabaseProvider.backToUIThread(coroutineScope) {
-            mViewMvc?.bindSession(mSessionPresenter)
-        }
     }
 
     @Subscribe(threadMode = ThreadMode.MAIN)
     fun onMessageEvent(event: NewMeasurementEvent) {
+        reloadSession()
+
         if (event.sensorName == mSessionPresenter?.selectedStream?.sensorName) {
             val location = LocationHelper.lastLocation()
             val measurement = Measurement(event, location?.latitude , location?.longitude)
 
-            mViewMvc?.addMeasurement(measurement)
+            mViewMvc?.addMeasurement(mSessionPresenter, measurement)
+        }
+    }
+
+    private fun reloadSession() {
+        val sessionUUID = mSessionPresenter?.sessionUUID ?: return
+
+        DatabaseProvider.runQuery { coroutineScope ->
+            val dbSession = mSessionsViewModel.reloadSession(sessionUUID)
+            dbSession?.let {
+                val session = Session(dbSession)
+                mSessionPresenter?.session = session
+
+                var selectedSensorName = mSessionPresenter.initialSensorName
+                if (mSessionPresenter.selectedStream != null) {
+                    selectedSensorName = mSessionPresenter.selectedStream!!.sensorName
+                }
+
+                val measurementStream =
+                    session.streams.firstOrNull { it.sensorName == selectedSensorName }
+                mSessionPresenter.selectedStream = measurementStream
+
+                val sensorThresholds = mSessionsViewModel.findOrCreateSensorThresholds(session)
+                mSessionPresenter.setSensorThresholds(sensorThresholds)
+            }
         }
     }
