+++ app/src/main/java/io/lunarlogic/aircasting/screens/session_view/SessionDetailsViewMvcImpl.kt
@@ -86,9 +86,11 @@ abstract class SessionDetailsViewMvcImpl: BaseObservableViewMvc<SessionDetailsVi
         mListener = null
     }
 
-    override fun addMeasurement(measurement: Measurement) {}
+    override fun addMeasurement(sessionPresenter: SessionPresenter?, measurement: Measurement) {
+        bindSession(sessionPresenter)
+    }
 
-    override fun bindSession(sessionPresenter: SessionPresenter?) {
+    protected open fun bindSession(sessionPresenter: SessionPresenter?) {
         mSessionPresenter = sessionPresenter
 
         if (sessionPresenter?.selectedStream?.measurements?.isNotEmpty() == true) { //TODO: check if works fine after Marysia's changes with memory leaks!
